UNIS-MUNDI-ICO
UNIS-MUNDI is a place where users can acquire and customize their personal Meta-Space in order to establish and offer their own content within their virtual estate. The LAND is the digital piece of a virtual estate in UNIS MUNDI metaverse registered as NFT ( BEP-721) and is de facto a scarce property. The LAND is the BEP-721 NFT that can be withdrawn to the wallet address of the user, and can be directly deposited to the inventory of the user account in UNIS MUNDI.app (Check the metamask wallet to ensure that the wallet address is correctly linked . By the way, check out my awesome work.